Start of streaming.
I want to be able to stream the Candidates. For large term
counts, there are tons of Candidates, however, the vast
majority will be filtered out in the first iteration. This means
that using a generator / streaming pattern should save a ton of
ram and making large term counts possible.

This is my first attempt. It works, but each Filter is evaluated
in order for the whole dataset, so the memory profile is still
substantial, albeit way better than before. A better
implementation would use a slice of channels, one for each
Filter. That way data can pass through the entire Filter chain.