Filtering semi integrated
It works, but for now I had to sacrifice some functionality.
Instead of generating a minimal number of problems needed to
make the operator sequence unqiue, the program just generates
n problems. This is largely due to streams generally not
tracking the number of items in them. Going to have to write
some ugly code to fix this.

Upside: I can now to 14 term problems, which is an upgrade.